Only use fibers if caller uses fibers.
Currently (without this change), callbacks passed to .then() will always run in a fiber when this package is active.

This is useful for Meteor code where many things must run in fibers. However, it is problematic for non-Meteor code that doesn't expect fibers. The main problem is performance: creating and switching fibers has a cost. Fiber-switching in particular trashes the cache.

Traditionally, promise-heavy code (that doesn't expect fibers) tends to involve lots of chaining of short continuations using .then(). Unfortunately, this style maximizes the overhead of creating a fiber for every callback.

Code that *expects* fibers can reduce the number of fibers created by taking advantage of them, e.g. by using .await() rather than chaining. But code which is not Meteor-specific cannot assume fibers.

The result is that promise-heavy code which happens to live in the same process as Meteor code will suffer severe performance penalties. Because the global definition of "then" is being overridden, it applies throughout the process, even to code that doesn't need or want fibers.

The best solution to this would have been never to have introduced this monkey-patch in the first place. Instead, a new method, e.g. `thenInFiber()`, should have been introduced.

However, it is probably too late for that now.

The opposite solution does *not* work: we cannot introduce a `thenNoFiber()` method because code which wasn't written for Meteor in the first place cannot be expected to use this.

As a compromise, this patch implements the rule: If the code which called `.then()` is in a fiber, then the continuations will be run in a fiber. If not, they won't.

This should have minimal impact on any Meteor code which expects `.then()` to run in a fiber, because the caller will also normally be in a fiber. Meanwhile, non-Meteor code can be kept out of fibers as long as developers are careful to schedule it outside of any fiber in the first place.

I've also added the methods `thenInFiber()` and `thenNotInFiber()` to help high-level code be explicit about pushing things in and out of fibers.

There's some chance this change will break existing code which explicitly expected .then() to add fibers when the caller wasn't in a fiber. I can't see a good way to solve the problem here without such breakage. Hopefully it is rare.

Fixes #9